requirements.txt
numpy
Keras=2.2.4
pandas==0.19.2
tensorflow_gpu==1.1.0
PyYAML